Re-work Status to be differentiated by OS
