+++ cli/Valet/Os/Linux.php
@@ -3,6 +3,9 @@
 namespace Valet\Os;
 
 use Valet\Os\Linux\Apt;
+use Valet\Os\Linux\LinuxStatus;
+use Valet\Status;
+
 use function Valet\resolve;
 
 class Linux extends Os
@@ -20,5 +23,10 @@ class Linux extends Os
     public function logDir(): string
     {
         return '/log';
+    }
+
+    public function status(): Status
+    {
+        return resolve(LinuxStatus::class);
     }
 }
