+++ tests/MacStatusTest.php
@@ -1,12 +1,16 @@
 <?php
 
 use Valet\CommandLine;
+use Valet\Os\Installer;
+use Valet\Os\Linux\LinuxStatus;
+use Valet\Os\Mac\Brew;
+use Valet\Os\Mac\MacStatus;
 use function Valet\resolve;
 use Valet\Status;
 use function Valet\swap;
 use function Valet\user;
 
-class StatusTest extends Yoast\PHPUnitPolyfills\TestCases\TestCase
+class MacStatusTest extends Yoast\PHPUnitPolyfills\TestCases\TestCase
 {
     use UsesNullWriter;
 
@@ -15,6 +19,7 @@ class StatusTest extends Yoast\PHPUnitPolyfills\TestCases\TestCase
         $_SERVER['SUDO_USER'] = user();
 
         $this->setNullWriter();
+        swap(Installer::class, resolve(Brew::class));
     }
 
     public function tear_down()
@@ -24,7 +29,10 @@ class StatusTest extends Yoast\PHPUnitPolyfills\TestCases\TestCase
 
     public function test_status_can_be_resolved_from_container()
     {
-        $this->assertInstanceOf(Status::class, resolve(Status::class));
+        $this->assertInstanceOf(MacStatus::class, resolve(MacStatus::class));
+
+        swap(Status::class, resolve(MacStatus::class));
+        $this->assertInstanceOf(MacStatus::class, resolve(Status::class));
     }
 
     public function test_it_checks_if_brew_services_are_running()
@@ -35,7 +43,7 @@ class StatusTest extends Yoast\PHPUnitPolyfills\TestCases\TestCase
 
         swap(CommandLine::class, $cli);
 
-        $status = resolve(Status::class);
+        $status = resolve(MacStatus::class);
 
         $this->assertTrue($status->isBrewServiceRunning('nginx'));
         $this->assertTrue($status->isBrewServiceRunning('php'));
@@ -49,7 +57,7 @@ class StatusTest extends Yoast\PHPUnitPolyfills\TestCases\TestCase
 
         swap(CommandLine::class, $cli);
 
-        $status = resolve(Status::class);
+        $status = resolve(MacStatus::class);
 
         $this->assertTrue($status->isBrewServiceRunning('nginx'));
         $this->assertTrue($status->isBrewServiceRunning('php', exactMatch: false));
